backend/requirements.txt
# FastAPI Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Request handling
python-multipart==0.0.6
httpx==0.26.0

# Image processing
pillow==10.2.0

# Firebase
firebase-admin==6.4.0

# Google Drive API
google-api-python-client==2.111.0
google-auth==2.25.2

# Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Testing (dev)
pytest==8.0.0
pytest-asyncio==0.23.3

